- adding extra vertex.rs - adding basicColorVertex struct
