+++ src/main.rs
@@ -2,6 +2,4 @@ use rui::run;
 
 fn main() {
     pollster::block_on(run());
-
-
 }
