+++ src/shapes.rs
@@ -1,3 +1,4 @@
 pub mod shape;
 pub mod quad;
 pub mod oval;
+pub mod vertex;
