+++ src/shapes/vertex.rs
@@ -1,0 +1,34 @@
+use wgpu::{BufferAddress, VertexBufferLayout};
+
+pub trait Vertex {
+    fn get_descriptor<'a>() -> VertexBufferLayout<'a>;
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
+pub struct BasicColorVertex {
+    pub position: [f32; 3],
+    pub color: [f32; 3],
+}
+
+impl Vertex for BasicColorVertex {
+    fn get_descriptor<'a>() -> VertexBufferLayout<'a> {
+        use std::mem;
+        VertexBufferLayout {
+            array_stride: mem::size_of::<BasicColorVertex>() as BufferAddress,
+            step_mode: wgpu::VertexStepMode::Vertex,
+            attributes: &[
+                wgpu::VertexAttribute {
+                    offset: 0,
+                    shader_location: 0,
+                    format: wgpu::VertexFormat::Float32x3,
+                },
+                wgpu::VertexAttribute {
+                    offset: mem::size_of::<[f32; 3]>() as BufferAddress,
+                    shader_location: 1,
+                    format: wgpu::VertexFormat::Float32x3,
+                }
+            ],
+        }
+    }
+}
